fix: qemu: uninitialized value in multiplication
A generated VM with default values does not set the memory key in the
configuration. Hence the size of the state file for a suspend had only
the default size of the state itself and not in addition twice the
configured memory.

The patch uses the static defaults from the JSON schema if the memory
key is not set.

Signed-off-by: Alwin Antreich <a.antreich@proxmox.com>
(cherry picked from commit 22ea69ca65b14bf0caf64dd0daf5c6bff675edf8)
Signed-off-by: Oguz Bektas <o.bektas@proxmox.com>